+++ lib/macos_ui.dart
@@ -14,6 +14,9 @@
 
 library macos_ui;
 
+export 'package:macos_window_utils/macos/ns_window_delegate.dart';
+export 'package:macos_window_utils/macos_window_utils.dart';
+
 export 'src/buttons/back_button.dart';
 export 'src/buttons/checkbox.dart';
 export 'src/buttons/disclosure_button.dart';
@@ -29,6 +32,7 @@ export 'src/buttons/toolbar/toolbar_icon_button.dart';
 export 'src/buttons/toolbar/toolbar_overflow_button.dart';
 export 'src/buttons/toolbar/toolbar_pulldown_button.dart';
 export 'src/dialogs/macos_alert_dialog.dart';
+export 'src/fields/autocomplete_field.dart';
 export 'src/fields/search_field.dart';
 export 'src/fields/text_field.dart';
 export 'src/icon/image_icon.dart';
@@ -37,7 +41,6 @@ export 'src/indicators/capacity_indicators.dart';
 export 'src/indicators/progress_indicators.dart';
 export 'src/indicators/rating_indicator.dart';
 export 'src/indicators/relevance_indicator.dart';
-export 'src/layout/scrollbar.dart';
 export 'src/indicators/slider.dart';
 export 'src/labels/label.dart';
 export 'src/labels/tooltip.dart';
@@ -45,6 +48,7 @@ export 'src/layout/content_area.dart';
 export 'src/layout/macos_list_tile.dart';
 export 'src/layout/resizable_pane.dart';
 export 'src/layout/scaffold.dart';
+export 'src/layout/scrollbar.dart';
 export 'src/layout/sidebar/sidebar.dart';
 export 'src/layout/sidebar/sidebar_item.dart';
 export 'src/layout/sidebar/sidebar_items.dart';
@@ -60,12 +64,15 @@ export 'src/layout/toolbar/toolbar_overflow_menu.dart';
 export 'src/layout/toolbar/toolbar_overflow_menu_item.dart';
 export 'src/layout/toolbar/toolbar_popup.dart';
 export 'src/layout/toolbar/toolbar_spacer.dart';
+export 'src/layout/wallpaper_tinted_area.dart';
 export 'src/layout/window.dart';
 export 'src/macos_app.dart';
+export 'src/macos_window_utils_config.dart';
 export 'src/selectors/color_well.dart';
 export 'src/selectors/date_picker.dart';
 export 'src/selectors/time_picker.dart';
 export 'src/sheets/macos_sheet.dart';
+export 'src/theme/autocomplete_field_theme.dart';
 export 'src/theme/date_picker_theme.dart';
 export 'src/theme/help_button_theme.dart';
 export 'src/theme/icon_button_theme.dart';
@@ -78,12 +85,6 @@ export 'src/theme/popup_button_theme.dart';
 export 'src/theme/pulldown_button_theme.dart';
 export 'src/theme/push_button_theme.dart';
 export 'src/theme/scrollbar_theme.dart';
-export 'src/theme/search_field_theme.dart';
 export 'src/theme/time_picker_theme.dart';
 export 'src/theme/tooltip_theme.dart';
 export 'src/theme/typography.dart';
-export 'src/layout/wallpaper_tinted_area.dart';
-export 'src/macos_window_utils_config.dart';
-
-export 'package:macos_window_utils/macos_window_utils.dart';
-export 'package:macos_window_utils/macos/ns_window_delegate.dart';
