+++ test/theme/autocomplete_field_theme_test.dart
@@ -4,41 +4,41 @@ import 'package:macos_ui/macos_ui.dart';
 import 'package:macos_ui/src/library.dart';
 
 void main() {
-  group('MacosSearchField theme tests', () {
+  group('MacosAutoCompleteField theme tests', () {
     test('lerps from light to dark', () {
-      final actual = MacosSearchFieldThemeData.lerp(
-        _macosSearchFieldTheme,
-        _macosSearchFieldThemeDark,
+      final actual = MacosAutoCompleteFieldThemeData.lerp(
+        _macosAutoCompleteFieldTheme,
+        _macosAutoCompleteFieldThemeDark,
         1,
       );
 
-      expect(actual, _macosSearchFieldThemeDark);
+      expect(actual, _macosAutoCompleteFieldThemeDark);
     });
 
     test('lerps from dark to light', () {
-      final actual = MacosSearchFieldThemeData.lerp(
-        _macosSearchFieldThemeDark,
-        _macosSearchFieldTheme,
+      final actual = MacosAutoCompleteFieldThemeData.lerp(
+        _macosAutoCompleteFieldThemeDark,
+        _macosAutoCompleteFieldTheme,
         1,
       );
 
-      expect(actual, _macosSearchFieldTheme);
+      expect(actual, _macosAutoCompleteFieldTheme);
     });
 
     test('copyWith, hashCode, ==', () {
       expect(
-        const MacosSearchFieldThemeData(),
-        const MacosSearchFieldThemeData().copyWith(),
+        const MacosAutoCompleteFieldThemeData(),
+        const MacosAutoCompleteFieldThemeData().copyWith(),
       );
       expect(
-        const MacosSearchFieldThemeData().hashCode,
-        const MacosSearchFieldThemeData().copyWith().hashCode,
+        const MacosAutoCompleteFieldThemeData().hashCode,
+        const MacosAutoCompleteFieldThemeData().copyWith().hashCode,
       );
     });
 
     testWidgets('debugFillProperties', (tester) async {
       final builder = DiagnosticPropertiesBuilder();
-      MacosSearchFieldThemeData(
+      MacosAutoCompleteFieldThemeData(
         highlightColor: CupertinoColors.activeBlue.color,
         resultsBackgroundColor: const Color.fromRGBO(242, 242, 247, 1),
       ).debugFillProperties(builder);
@@ -62,14 +62,13 @@ void main() {
       await tester.pumpWidget(
         MacosApp(
           home: MacosWindow(
-            disableWallpaperTinting: true,
             child: MacosScaffold(
               children: [
                 ContentArea(
                   builder: (context, _) {
                     capturedContext = context;
                     return const Center(
-                      child: MacosSearchField(),
+                      child: MacosAutoCompleteField(),
                     );
                   },
                 ),
@@ -79,19 +78,19 @@ void main() {
         ),
       );
 
-      final theme = MacosSearchFieldTheme.of(capturedContext);
+      final theme = MacosAutoCompleteFieldTheme.of(capturedContext);
       expect(theme.highlightColor, const Color(0xff007aff));
       expect(theme.resultsBackgroundColor, const Color(0xfff2f2f7));
     });
   });
 }
 
-final _macosSearchFieldTheme = MacosSearchFieldThemeData(
+final _macosAutoCompleteFieldTheme = MacosAutoCompleteFieldThemeData(
   highlightColor: CupertinoColors.activeBlue.color,
   resultsBackgroundColor: const Color.fromRGBO(242, 242, 247, 1),
 );
 
-final _macosSearchFieldThemeDark = MacosSearchFieldThemeData(
+final _macosAutoCompleteFieldThemeDark = MacosAutoCompleteFieldThemeData(
   highlightColor: CupertinoColors.activeBlue.darkColor,
   resultsBackgroundColor: const Color.fromRGBO(30, 30, 30, 1),
 );
